+++ src/app/(faculty)/faculty/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import getCurrentUser from "@/app/actions/getCurrentUser";
+import getCurrentUser from "@/actions/getCurrentUser";
 import FacultyDashboard from "@/components/faculty/dashboard/faculty-dashboard";
 
 export const metadata = {
