+++ src/app/(student)/student/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import getCurrentUser from "@/app/actions/getCurrentUser";
+import getCurrentUser from "@/actions/getCurrentUser";
 import StudentDashboard from "@/components/student/dashboard/student-dashboard";
 
 export const metadata = {
