+++ src/components/student/dashboard/student-dashboard.tsx
@@ -1,6 +1,6 @@
 import Heading from "@/components/ui/heading";
 import DashboardCalendar from "@/components/dashboard-calendar";
-import getCurrentUser from "@/app/actions/getCurrentUser";
+import getCurrentUser from "@/actions/getCurrentUser";
 
 export default async function StudentDashboard() {
   const currentUser = await getCurrentUser();
